fix all project

+++ includes/algorithms/matrix.h
@@ -48,11 +48,11 @@ void swapRows(matrix m, int i1, int i2);
 
 void swapColumns(matrix m, int j1, int j2);
 
-void insertionSortRowsMatrixByRowCriteria(matrix m, long long (*criteria)(const int *, size_t));
+void insertionSortRowsMatrixByRowCriteria(matrix m, int (*criteria)(const int *, size_t));
 
 void insertionSortRowsMatrixByRowCriteriaD(matrix m, double (*criteria)(const int *, size_t));
 
-void insertionSortColsMatrixByColCriteria(matrix m, long long (*criteria)(const int *, size_t));
+void insertionSortColsMatrixByColCriteria(matrix m, int (*criteria)(const int *, size_t));
 
 bool isSquareMatrix(matrix m);
 
@@ -78,7 +78,7 @@ matrix mulMatrices(matrix m1, matrix m2);
 
 position getLeftMinPositionElement(matrix m);
 
-bool hasAllSortByRows(matrix m, bool(*condition)(long long, long long));
+bool hasAllSortByRows(matrix m, bool(*condition)(int, int));
 
 unsigned getCounterZeroRows(matrix m);
